Extract profile componenets into directories and update file paths
