+++ client/src/App.js
@@ -2,7 +2,7 @@ import React, { useEffect } from "react"
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import Today from './components/day/Today'
 import Cover from "./components/cover/Cover"
-import EditProfile from "./components/EditProfile";
+import EditProfile from "./components/profile/EditProfile";
 import AppState from './context/application/AppState'
 import AuthState from './context/authentication/AuthState'
 import setAuthToken from "./utils/setAuthToken"
