+++ client/src/components/profile/EditProfile.js
@@ -1,11 +1,11 @@
 import React, { useContext, useEffect } from 'react'
-import Alert from "../components/Alert"
+import Alert from "../Alert"
 import { Link } from 'react-router-dom'
 import ArrowBackIosIcon from '@material-ui/icons/ArrowBackIos';
 import PersonIcon from '@material-ui/icons/Person';
-import AuthContext from "../context/authentication/AuthContext"
+import AuthContext from "../../context/authentication/AuthContext"
 
-import pageFlip from "../sounds/page-flip.wav"
+import pageFlip from "../../sounds/page-flip.wav"
 import EditInfo from './EditInfo';
 import PasswordField from './PasswordField';
